create utils/contansts
